requirements.txt
feedparser==6.0.10
requests==2.32.3
Jinja2==3.1.4
python-dateutil==2.9.0.post0
openai==1.52.2